Migrate Mutation.addAlert
